Extract EnSight binary readers into detail/ensight_io
